Provide node Streams-based API
And endure the horror that is Node streaming...

The aim of this change is to to make the Hoard API deal in
native Node streams particularly byte streams (also message objects
stream in various places) for sending file data (body) and make it so
consumers do not need to know or care about things like the order of
messages on the wire, since it had become non-obvious how to use the
various helper functions to actually send stuff. The lower-level GRPC
services are available on the Hoard client but should not be needed all
top level methods accept Stream-like inputs where applicable (Streams,
iterables, Buffers, Uint8Arrays) and return streams.

Some things to note:

- `readBytes` and `readAll` helpers are promise-returning helpers to read
the values for streams. `read` is a reducer that can do things like read
partial streams.

- The hoard client is now an actual class (like it probably always should
have been) so is created with `new`.

- All bufferring/chunking is handled transparently by the client so just
pass a byte-mode stream where applicable.

- GRPC messages like Header and Spec that are required to being a stream
and be sent only once are now taken as separate arguments and the client
handles multiplexing them onto the underlying GRPC stream.

The biggest issue developing this was the (surprising) lack of decent
primitives for composing streams in a way that doesn't complete mess up
error handling. I tried stream.pipeline but it doesn't give you a duplex
stream, and I nearly used pumpify, but the hideous dependency graph of
tiny packages with dubious legacy content put me off, as well as them
not properly inheriting from Stream so I have implemented my own
`pipeline` that seems to work. I hope with node v15 there will not be
too may additional corner cases but we may find there are some issues
with leaks or something like that, but it feels like it is preferable to
just fix them when they come up.

Signed-off-by: Silas Davis <silas@monax.io>